preprocessing/requirements.txt
#plasmidpicker
click
biopython
numpy
tqdm
h5py